# my-page-api/src/main/resources/openapi/api.yaml
openapi: 3.0.3
info:
  title: JPro My Page API
  description: API for JPro My Page
  version: 1.0.0
paths:
...
  /booking:
    get:
      tags:
        - Booking
      operationId: getBookings
      summary: Get bookings
      description: Returns bookings within a date range
      parameters:
        - in: query
          name: startDate
          required: true
          schema:
            type: string
            format: date
        - in: query
          name: endDate
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: List of bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Booking'
...
  /booking/{bookingId}:
    delete:
      tags:
        - Booking
      operationId: deleteBooking
      summary: Delete booking
      description: Delete a booking by ID
      parameters:
        - name: bookingId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Booking deleted
    patch:
      tags:
        - Booking
      operationId: updateBooking
      summary: Update booking
      description: Update a booking by ID
      parameters:
        - name: bookingId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookingUpdate'
      responses:
        '200':
          description: Booking updated
...
  /booking/admin/post:
    post:
      tags:
        - Booking
      operationId: adminCreateBooking
      summary: Admin creates a booking for a user
      description: Admin creates a new booking on behalf of a specific user
      parameters:
        - name: bookingOwnerName
          in: query
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBooking'
      responses:
        '201':
          description: Booking created successfully
          content:
            text/plain:
              schema:
                type: string
        '403':
          description: User not found
  /budget/{employeeNumber}:
    get:
      tags:
        - Budget
      operationId: getBudgetsForEmployee
      summary: Get budgets for employee
      description: Returns all budgets for a specific employee (admin only)
      parameters:
        - name: employeeNumber
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: List of budgets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BudgetDetail'
  /budget/{budgetId}/posts:
    post:
      tags:
        - Budget
      operationId: createBudgetPost
      summary: Create budget post
      description: Create a new post for a budget
      parameters:
        - name: budgetId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePost'
      responses:
        '201':
          description: Post created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '400':
          description: Bad request
        '403':
          description: Forbidden
  /budget/posts/{postId}:
    delete:
      tags:
        - Budget
      operationId: deleteBudgetPost
      summary: Delete budget post
      description: Delete a budget post by ID
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Post deleted
        '403':
          description: Forbidden
        '404':
          description: Not found
    patch:
      tags:
        - Budget
      operationId: updateBudgetPost
      summary: Update budget post
      description: Update an existing budget post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePost'
      responses:
        '200':
          description: Post updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '403':
          description: Forbidden
        '404':
          description: Not found
...
components:
  schemas:
...
    Booking:
      type: object
      properties:
        id:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        apartment:
          $ref: '#/components/schemas/Apartment'
        employeeName:
          type: string
        isPending:
          type: boolean
      required:
        - id
        - startDate
        - endDate
        - apartment
    Apartment:
      type: object
      properties:
        id:
          type: integer
          format: int64
        cabin_name:
          type: string
        sort_order:
          type: integer
      required:
        - id
        - cabin_name
    BookingUpdate:
      type: object
      properties:
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        apartmentId:
          type: integer
          format: int64
    BudgetDetail:
      type: object
      description: Detailed budget information including posts
      properties:
        id:
          type: integer
          format: int64
        posts:
          type: array
          items:
            $ref: '#/components/schemas/Post'
        budgetType:
          type: object
          properties:
            id:
              type: integer
              format: int64
            name:
              type: string
            startAmount:
              type: number
        startDate:
          type: string
          format: date
        startAmount:
          type: number
        hours:
          type: array
          items:
            type: object
      required:
        - id
        - posts
        - budgetType
        - startDate
        - startAmount
    CreatePost:
      type: object
      properties:
        date:
          type: string
          format: date
        description:
          type: string
        amountIncMva:
          type: number
        amountExMva:
          type: number
        expense:
          type: boolean
      required:
        - date
        - amountIncMva
        - amountExMva
        - expense
    Post:
      type: object
      properties:
        id:
          type: integer
          format: int64
        date:
          type: string
          format: date
        description:
          type: string
        amountIncMva:
          type: number
        amountExMva:
          type: number
        documentNumber:
          type: string
        dateOfPayment:
          type: string
          format: date
        dateOfDeduction:
          type: string
          format: date
        expense:
          type: boolean
        locked:
          type: boolean
        createdDate:
          type: string
          format: date-time
        lastModifiedDate:
          type: string
          format: date-time
        createdBy:
          type: string
      required:
        - id
        - date
        - expense
        - locked
    UpdatePost:
      type: object
      properties:
        date:
          type: string
          format: date
        description:
          type: string
        amountExMva:
          type: number
...
    CreateBooking:
      type: object
      properties:
        apartmentID:
          type: integer
          format: int64
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
      required:
        - apartmentID
        - startDate
        - endDate
...